+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Unreleased
 
+* [null]
+
+## 0.0.33 (June 14, 2018)
+
 * exclude local node_modules
-* disable fork-ts to show the error in browser
+* disable fork-ts to show errors in browser
 * 支持 Shell 生命周期 **before** 修改 Webpack.options
 
 ## 0.0.32 (June 11, 2018)